Fix keyboard navigation and deploy SolidJS build
- Navigation now respects display sort order
- Focus maintained during j/k arrow key navigation
- Remove old vanilla JS, deploy built SolidJS app
- Fix unused imports in test files
