+++ web-solid/src/test/keyboard.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect, beforeEach } from 'vitest';
+import { describe, it, expect } from 'vitest';
 
 describe('Keyboard shortcuts', () => {
   // Test the switch statement logic directly
