+++ web/assets/index-D0VPOhP_.js
@@ -1,0 +1,1 @@
+(function(){const t=document.createElement("link").relList;if(t&&t.supports&&t.supports("modulepreload"))return;for(const r of document.querySelectorAll('link[rel="modulepreload"]'))o(r);new MutationObserver(r=>{for(const s of r)if(s.type==="childList")for(const l of s.addedNodes)l.tagName==="LINK"&&l.rel==="modulepreload"&&o(l)}).observe(document,{childList:!0,subtree:!0});function n(r){const s={};return r.integrity&&(s.integrity=r.integrity),r.referrerPolicy&&(s.referrerPolicy=r.referrerPolicy),r.crossOrigin==="use-credentials"?s.credentials="include":r.crossOrigin==="anonymous"?s.credentials="omit":s.credentials="same-origin",s}function o(r){if(r.ep)return;r.ep=!0;const s=n(r);fetch(r.href,s)}})();const st=!1,rt=(e,t)=>e===t,K=Symbol("solid-proxy"),me=Symbol("solid-track"),ue={equals:rt};let We=Ge;const z=1,fe=2,Ve={owned:null,cleanups:null,context:null,owner:null};var E=null;let we=null,ot=null,C=null,D=null,L=null,$e=0;function ce(e,t){const n=C,o=E,r=e.length===0,s=t===void 0?o:t,l=r?Ve:{owned:null,cleanups:null,context:s?s.context:null,owner:s},i=r?e:()=>e(()=>U(()=>ne(l)));E=l,C=null;try{return Q(i,!0)}finally{C=n,E=o}}function de(e,t){t=t?Object.assign({},ue,t):ue;const n={value:e,observers:null,observerSlots:null,comparator:t.equals||void 0},o=r=>(typeof r=="function"&&(r=r(n.value)),Ue(n,r));return[Ke.bind(n),o]}function m(e,t,n){const o=Ee(e,t,!1,z);oe(o)}function ze(e,t,n){We=ut;const o=Ee(e,t,!1,z);o.user=!0,L?L.push(o):oe(o)}function q(e,t,n){n=n?Object.assign({},ue,n):ue;const o=Ee(e,t,!0,0);return o.observers=null,o.observerSlots=null,o.comparator=n.equals||void 0,oe(o),Ke.bind(o)}function lt(e){return Q(e,!1)}function U(e){if(C===null)return e();const t=C;C=null;try{return e()}finally{C=t}}function it(e){ze(()=>U(e))}function He(e){return E===null||(E.cleanups===null?E.cleanups=[e]:E.cleanups.push(e)),e}function ke(){return C}function Ke(){if(this.sources&&this.state)if(this.state===z)oe(this);else{const e=D;D=null,Q(()=>_e(this),!1),D=e}if(C){const e=this.observers?this.observers.length:0;C.sources?(C.sources.push(this),C.sourceSlots.push(e)):(C.sources=[this],C.sourceSlots=[e]),this.observers?(this.observers.push(C),this.observerSlots.push(C.sources.length-1)):(this.observers=[C],this.observerSlots=[C.sources.length-1])}return this.value}function Ue(e,t,n){let o=e.value;return(!e.comparator||!e.comparator(o,t))&&(e.value=t,e.observers&&e.observers.length&&Q(()=>{for(let r=0;r<e.observers.length;r+=1){const s=e.observers[r],l=we&&we.running;l&&we.disposed.has(s),(l?!s.tState:!s.state)&&(s.pure?D.push(s):L.push(s),s.observers&&Je(s)),l||(s.state=z)}if(D.length>1e6)throw D=[],new Error},!1)),t}function oe(e){if(!e.fn)return;ne(e);const t=$e;at(e,e.value,t)}function at(e,t,n){let o;const r=E,s=C;C=E=e;try{o=e.fn(t)}catch(l){return e.pure&&(e.state=z,e.owned&&e.owned.forEach(ne),e.owned=null),e.updatedAt=n+1,Ye(l)}finally{C=s,E=r}(!e.updatedAt||e.updatedAt<=n)&&(e.updatedAt!=null&&"observers"in e?Ue(e,o):e.value=o,e.updatedAt=n)}function Ee(e,t,n,o=z,r){const s={fn:e,state:o,updatedAt:null,owned:null,sources:null,sourceSlots:null,cleanups:null,value:t,owner:E,context:E?E.context:null,pure:n};return E===null||E!==Ve&&(E.owned?E.owned.push(s):E.owned=[s]),s}function he(e){if(e.state===0)return;if(e.state===fe)return _e(e);if(e.suspense&&U(e.suspense.inFallback))return e.suspense.effects.push(e);const t=[e];for(;(e=e.owner)&&(!e.updatedAt||e.updatedAt<$e);)e.state&&t.push(e);for(let n=t.length-1;n>=0;n--)if(e=t[n],e.state===z)oe(e);else if(e.state===fe){const o=D;D=null,Q(()=>_e(e,t[0]),!1),D=o}}function Q(e,t){if(D)return e();let n=!1;t||(D=[]),L?n=!0:L=[],$e++;try{const o=e();return ct(n),o}catch(o){n||(L=null),D=null,Ye(o)}}function ct(e){if(D&&(Ge(D),D=null),e)return;const t=L;L=null,t.length&&Q(()=>We(t),!1)}function Ge(e){for(let t=0;t<e.length;t++)he(e[t])}function ut(e){let t,n=0;for(t=0;t<e.length;t++){const o=e[t];o.user?e[n++]=o:he(o)}for(t=0;t<n;t++)he(e[t])}function _e(e,t){e.state=0;for(let n=0;n<e.sources.length;n+=1){const o=e.sources[n];if(o.sources){const r=o.state;r===z?o!==t&&(!o.updatedAt||o.updatedAt<$e)&&he(o):r===fe&&_e(o,t)}}}function Je(e){for(let t=0;t<e.observers.length;t+=1){const n=e.observers[t];n.state||(n.state=fe,n.pure?D.push(n):L.push(n),n.observers&&Je(n))}}function ne(e){let t;if(e.sources)for(;e.sources.length;){const n=e.sources.pop(),o=e.sourceSlots.pop(),r=n.observers;if(r&&r.length){const s=r.pop(),l=n.observerSlots.pop();o<r.length&&(s.sourceSlots[l]=o,r[o]=s,n.observerSlots[o]=l)}}if(e.tOwned){for(t=e.tOwned.length-1;t>=0;t--)ne(e.tOwned[t]);delete e.tOwned}if(e.owned){for(t=e.owned.length-1;t>=0;t--)ne(e.owned[t]);e.owned=null}if(e.cleanups){for(t=e.cleanups.length-1;t>=0;t--)e.cleanups[t]();e.cleanups=null}e.state=0}function ft(e){return e instanceof Error?e:new Error(typeof e=="string"?e:"Unknown error",{cause:e})}function Ye(e,t=E){throw ft(e)}const dt=Symbol("fallback");function Te(e){for(let t=0;t<e.length;t++)e[t]()}function ht(e,t,n={}){let o=[],r=[],s=[],l=0,i=t.length>1?[]:null;return He(()=>Te(s)),()=>{let a=e()||[],u=a.length,_,c;return a[me],U(()=>{let g,f,$,k,x,w,P,v,p;if(u===0)l!==0&&(Te(s),s=[],o=[],r=[],l=0,i&&(i=[])),n.fallback&&(o=[dt],r[0]=ce(O=>(s[0]=O,n.fallback())),l=1);else if(l===0){for(r=new Array(u),c=0;c<u;c++)o[c]=a[c],r[c]=ce(y);l=u}else{for($=new Array(u),k=new Array(u),i&&(x=new Array(u)),w=0,P=Math.min(l,u);w<P&&o[w]===a[w];w++);for(P=l-1,v=u-1;P>=w&&v>=w&&o[P]===a[v];P--,v--)$[v]=r[P],k[v]=s[P],i&&(x[v]=i[P]);for(g=new Map,f=new Array(v+1),c=v;c>=w;c--)p=a[c],_=g.get(p),f[c]=_===void 0?-1:_,g.set(p,c);for(_=w;_<=P;_++)p=o[_],c=g.get(p),c!==void 0&&c!==-1?($[c]=r[_],k[c]=s[_],i&&(x[c]=i[_]),c=f[c],g.set(p,c)):s[_]();for(c=w;c<u;c++)c in $?(r[c]=$[c],s[c]=k[c],i&&(i[c]=x[c],i[c](c))):r[c]=ce(y);r=r.slice(0,l=u),o=a.slice(0)}return r});function y(g){if(s[c]=g,i){const[f,$]=de(c);return i[c]=$,t(a[c],f)}return t(a[c])}}}function T(e,t){return U(()=>e(t||{}))}const _t=e=>`Stale read from <${e}>.`;function xe(e){const t="fallback"in e&&{fallback:()=>e.fallback};return q(ht(()=>e.each,e.children,t||void 0))}function pe(e){const t=e.keyed,n=q(()=>e.when,void 0,void 0),o=t?n:q(n,void 0,{equals:(r,s)=>!r==!s});return q(()=>{const r=o();if(r){const s=e.children;return typeof s=="function"&&s.length>0?U(()=>s(t?r:()=>{if(!U(o))throw _t("Show");return n()})):s}return e.fallback},void 0,void 0)}const Y=e=>q(()=>e());function pt(e,t,n){let o=n.length,r=t.length,s=o,l=0,i=0,a=t[r-1].nextSibling,u=null;for(;l<r||i<s;){if(t[l]===n[i]){l++,i++;continue}for(;t[r-1]===n[s-1];)r--,s--;if(r===l){const _=s<o?i?n[i-1].nextSibling:n[s-i]:a;for(;i<s;)e.insertBefore(n[i++],_)}else if(s===i)for(;l<r;)(!u||!u.has(t[l]))&&t[l].remove(),l++;else if(t[l]===n[s-1]&&n[i]===t[r-1]){const _=t[--r].nextSibling;e.insertBefore(n[i++],t[l++].nextSibling),e.insertBefore(n[--s],_),t[r]=n[s]}else{if(!u){u=new Map;let c=i;for(;c<s;)u.set(n[c],c++)}const _=u.get(t[l]);if(_!=null)if(i<_&&_<s){let c=l,y=1,g;for(;++c<r&&c<s&&!((g=u.get(t[c]))==null||g!==_+y);)y++;if(y>_-i){const f=t[l];for(;i<_;)e.insertBefore(n[i++],f)}else e.replaceChild(n[i++],t[l++])}else l++;else t[l++].remove()}}}const Ie="_$DX_DELEGATE";function vt(e,t,n,o={}){let r;return ce(s=>{r=s,t===document?e():b(t,e(),t.firstChild?null:void 0,n)},o.owner),()=>{r(),t.textContent=""}}function A(e,t,n,o){let r;const s=()=>{const i=document.createElement("template");return i.innerHTML=e,i.content.firstChild},l=()=>(r||(r=s())).cloneNode(!0);return l.cloneNode=l,l}function le(e,t=window.document){const n=t[Ie]||(t[Ie]=new Set);for(let o=0,r=e.length;o<r;o++){const s=e[o];n.has(s)||(n.add(s),t.addEventListener(s,bt))}}function d(e,t){t==null?e.removeAttribute("class"):e.className=t}function qe(e,t,n,o){Array.isArray(n)?(e[`$$${t}`]=n[0],e[`$$${t}Data`]=n[1]):e[`$$${t}`]=n}function b(e,t,n,o){if(n!==void 0&&!o&&(o=[]),typeof t!="function")return ve(e,t,o,n);m(r=>ve(e,t(),r,n),o)}function bt(e){let t=e.target;const n=`$$${e.type}`,o=e.target,r=e.currentTarget,s=a=>Object.defineProperty(e,"target",{configurable:!0,value:a}),l=()=>{const a=t[n];if(a&&!t.disabled){const u=t[`${n}Data`];if(u!==void 0?a.call(t,u,e):a.call(t,e),e.cancelBubble)return}return t.host&&typeof t.host!="string"&&!t.host._$host&&t.contains(e.target)&&s(t.host),!0},i=()=>{for(;l()&&(t=t._$host||t.parentNode||t.host););};if(Object.defineProperty(e,"currentTarget",{configurable:!0,get(){return t||document}}),e.composedPath){const a=e.composedPath();s(a[0]);for(let u=0;u<a.length-2&&(t=a[u],!!l());u++){if(t._$host){t=t._$host,i();break}if(t.parentNode===r)break}}else i();s(o)}function ve(e,t,n,o,r){for(;typeof n=="function";)n=n();if(t===n)return n;const s=typeof t,l=o!==void 0;if(e=l&&n[0]&&n[0].parentNode||e,s==="string"||s==="number"){if(s==="number"&&(t=t.toString(),t===n))return n;if(l){let i=n[0];i&&i.nodeType===3?i.data!==t&&(i.data=t):i=document.createTextNode(t),n=G(e,n,o,i)}else n!==""&&typeof n=="string"?n=e.firstChild.data=t:n=e.textContent=t}else if(t==null||s==="boolean")n=G(e,n,o);else{if(s==="function")return m(()=>{let i=t();for(;typeof i=="function";)i=i();n=ve(e,i,n,o)}),()=>n;if(Array.isArray(t)){const i=[],a=n&&Array.isArray(n);if(Pe(i,t,n,r))return m(()=>n=ve(e,i,n,o,!0)),()=>n;if(i.length===0){if(n=G(e,n,o),l)return n}else a?n.length===0?Ne(e,i,o):pt(e,n,i):(n&&G(e),Ne(e,i));n=i}else if(t.nodeType){if(Array.isArray(n)){if(l)return n=G(e,n,o,t);G(e,n,null,t)}else n==null||n===""||!e.firstChild?e.appendChild(t):e.replaceChild(t,e.firstChild);n=t}}return n}function Pe(e,t,n,o){let r=!1;for(let s=0,l=t.length;s<l;s++){let i=t[s],a=n&&n[e.length],u;if(!(i==null||i===!0||i===!1))if((u=typeof i)=="object"&&i.nodeType)e.push(i);else if(Array.isArray(i))r=Pe(e,i,a)||r;else if(u==="function")if(o){for(;typeof i=="function";)i=i();r=Pe(e,Array.isArray(i)?i:[i],Array.isArray(a)?a:[a])||r}else e.push(i),r=!0;else{const _=String(i);a&&a.nodeType===3&&a.data===_?e.push(a):e.push(document.createTextNode(_))}}return r}function Ne(e,t,n=null){for(let o=0,r=t.length;o<r;o++)e.insertBefore(t[o],n)}function G(e,t,n,o){if(n===void 0)return e.textContent="";const r=o||document.createTextNode("");if(t.length){let s=!1;for(let l=t.length-1;l>=0;l--){const i=t[l];if(r!==i){const a=i.parentNode===e;!s&&!l?a?e.replaceChild(r,i):e.insertBefore(r,n):a&&i.remove()}else s=!0}}else e.insertBefore(r,n);return[r]}const Ce=Symbol("store-raw"),X=Symbol("store-node"),R=Symbol("store-has"),Xe=Symbol("store-self");function Qe(e){let t=e[K];if(!t&&(Object.defineProperty(e,K,{value:t=new Proxy(e,$t)}),!Array.isArray(e))){const n=Object.keys(e),o=Object.getOwnPropertyDescriptors(e);for(let r=0,s=n.length;r<s;r++){const l=n[r];o[l].get&&Object.defineProperty(e,l,{enumerable:o[l].enumerable,get:o[l].get.bind(t)})}}return t}function be(e){let t;return e!=null&&typeof e=="object"&&(e[K]||!(t=Object.getPrototypeOf(e))||t===Object.prototype||Array.isArray(e))}function se(e,t=new Set){let n,o,r,s;if(n=e!=null&&e[Ce])return n;if(!be(e)||t.has(e))return e;if(Array.isArray(e)){Object.isFrozen(e)?e=e.slice(0):t.add(e);for(let l=0,i=e.length;l<i;l++)r=e[l],(o=se(r,t))!==r&&(e[l]=o)}else{Object.isFrozen(e)?e=Object.assign({},e):t.add(e);const l=Object.keys(e),i=Object.getOwnPropertyDescriptors(e);for(let a=0,u=l.length;a<u;a++)s=l[a],!i[s].get&&(r=e[s],(o=se(r,t))!==r&&(e[s]=o))}return e}function ge(e,t){let n=e[t];return n||Object.defineProperty(e,t,{value:n=Object.create(null)}),n}function re(e,t,n){if(e[t])return e[t];const[o,r]=de(n,{equals:!1,internal:!0});return o.$=r,e[t]=o}function gt(e,t){const n=Reflect.getOwnPropertyDescriptor(e,t);return!n||n.get||!n.configurable||t===K||t===X||(delete n.value,delete n.writable,n.get=()=>e[K][t]),n}function Ze(e){ke()&&re(ge(e,X),Xe)()}function yt(e){return Ze(e),Reflect.ownKeys(e)}const $t={get(e,t,n){if(t===Ce)return e;if(t===K)return n;if(t===me)return Ze(e),n;const o=ge(e,X),r=o[t];let s=r?r():e[t];if(t===X||t===R||t==="__proto__")return s;if(!r){const l=Object.getOwnPropertyDescriptor(e,t);ke()&&(typeof s!="function"||e.hasOwnProperty(t))&&!(l&&l.get)&&(s=re(o,t,s)())}return be(s)?Qe(s):s},has(e,t){return t===Ce||t===K||t===me||t===X||t===R||t==="__proto__"?!0:(ke()&&re(ge(e,R),t)(),t in e)},set(){return!0},deleteProperty(){return!0},ownKeys:yt,getOwnPropertyDescriptor:gt};function ye(e,t,n,o=!1){if(!o&&e[t]===n)return;const r=e[t],s=e.length;n===void 0?(delete e[t],e[R]&&e[R][t]&&r!==void 0&&e[R][t].$()):(e[t]=n,e[R]&&e[R][t]&&r===void 0&&e[R][t].$());let l=ge(e,X),i;if((i=re(l,t,r))&&i.$(()=>n),Array.isArray(e)&&e.length!==s){for(let a=e.length;a<s;a++)(i=l[a])&&i.$();(i=re(l,"length",s))&&i.$(e.length)}(i=l[Xe])&&i.$()}function et(e,t){const n=Object.keys(t);for(let o=0;o<n.length;o+=1){const r=n[o];ye(e,r,t[r])}}function wt(e,t){if(typeof t=="function"&&(t=t(e)),t=se(t),Array.isArray(t)){if(e===t)return;let n=0,o=t.length;for(;n<o;n++){const r=t[n];e[n]!==r&&ye(e,n,r)}ye(e,"length",o)}else et(e,t)}function te(e,t,n=[]){let o,r=e;if(t.length>1){o=t.shift();const l=typeof o,i=Array.isArray(e);if(Array.isArray(o)){for(let a=0;a<o.length;a++)te(e,[o[a]].concat(t),n);return}else if(i&&l==="function"){for(let a=0;a<e.length;a++)o(e[a],a)&&te(e,[a].concat(t),n);return}else if(i&&l==="object"){const{from:a=0,to:u=e.length-1,by:_=1}=o;for(let c=a;c<=u;c+=_)te(e,[c].concat(t),n);return}else if(t.length>1){te(e[o],t,[o].concat(n));return}r=e[o],n=[o].concat(n)}let s=t[0];typeof s=="function"&&(s=s(r,n),s===r)||o===void 0&&s==null||(s=se(s),o===void 0||be(r)&&be(s)&&!Array.isArray(s)?et(r,s):ye(e,o,s))}function St(...[e,t]){const n=se(e||{}),o=Array.isArray(n),r=Qe(n);function s(...l){lt(()=>{o&&l.length===1?wt(n,l[0]):te(n,l)})}return[r,s]}const tt="bearing-state";function mt(){try{const e=localStorage.getItem(tt);if(e){const t=JSON.parse(e);let n=t.currentView||"operational";return n==="worktrees"&&(n="operational"),(n==="issues"||n==="prs")&&(n="planning"),{...t,currentView:n}}}catch(e){console.warn("Failed to load persisted state:",e)}return{}}const F=mt(),[h,I]=St({currentView:F.currentView||"operational",projects:[],worktrees:[],plans:[],selectedProject:F.selectedProject||null,selectedWorktreeFolder:F.selectedWorktreeFolder||null,selectedPlanPath:F.selectedPlanPath||null,focusedPanel:"project-list",sortColumn:F.sortColumn||"default",sortDirection:F.sortDirection||"asc",planSortColumn:F.planSortColumn||"default",planSortDirection:F.planSortDirection||"asc"});ze(()=>{const e={currentView:h.currentView,selectedProject:h.selectedProject,selectedWorktreeFolder:h.selectedWorktreeFolder,selectedPlanPath:h.selectedPlanPath,sortColumn:h.sortColumn,sortDirection:h.sortDirection,planSortColumn:h.planSortColumn,planSortDirection:h.planSortDirection};localStorage.setItem(tt,JSON.stringify(e))});function Ae(e){I("currentView",e)}function kt(e){I("projects",e)}function Pt(e){I("worktrees",e)}function Ct(e){I("plans",e)}function At(e){I("selectedProject",e)}function jt(e){I("selectedWorktreeFolder",e)}function Et(e){I("selectedPlanPath",e)}function V(e){I("focusedPanel",e)}function xt(e){I("sortColumn",e)}function Re(e){I("sortDirection",e)}function Dt(e){I("planSortColumn",e)}function Be(e){I("planSortDirection",e)}function Ot(){return h.worktrees.find(e=>e.folder===h.selectedWorktreeFolder)}function Tt(){return h.plans.find(e=>e.path===h.selectedPlanPath)}function Le(e){const t=h.projects;if(t.length===0)return;const n=t.findIndex(r=>r.name===h.selectedProject);let o;n===-1?o=e==="down"?0:t.length-1:e==="down"?o=Math.min(n+1,t.length-1):o=Math.max(n-1,0),I("selectedProject",t[o].name)}function It(e,t,n){const o=[...e],r=n==="asc"?1:-1;if(t==="default"){const s={OPEN:0,DRAFT:1,MERGED:2,CLOSED:3};o.sort((l,i)=>{const a=s[l.prState||""]??4,u=s[i.prState||""]??4;return a!==u?(a-u)*r:l.dirty!==i.dirty?((i.dirty?1:0)-(l.dirty?1:0))*r:l.folder.localeCompare(i.folder)*r})}else if(t==="folder")o.sort((s,l)=>s.folder.localeCompare(l.folder)*r);else if(t==="branch")o.sort((s,l)=>s.branch.localeCompare(l.branch)*r);else if(t==="status")o.sort((s,l)=>{const i=s.dirty?0:s.unpushed>0?1:2,a=l.dirty?0:l.unpushed>0?1:2;return(i-a)*r});else if(t==="pr"){const s={OPEN:0,DRAFT:1,MERGED:2,CLOSED:3};o.sort((l,i)=>{const a=s[l.prState||""]??4,u=s[i.prState||""]??4;return(a-u)*r})}return o}function Fe(e){const t=h.worktrees.filter(s=>s.repo===h.selectedProject),n=It(t,h.sortColumn,h.sortDirection);if(n.length===0)return;const o=n.findIndex(s=>s.folder===h.selectedWorktreeFolder);let r;o===-1?r=e==="down"?0:n.length-1:e==="down"?r=Math.min(o+1,n.length-1):r=Math.max(o-1,0),I("selectedWorktreeFolder",n[r].folder)}function Me(e){const t=h.plans.filter(r=>r.project===h.selectedProject);if(t.length===0)return;const n=t.findIndex(r=>r.path===h.selectedPlanPath);let o;n===-1?o=e==="down"?0:t.length-1:e==="down"?o=Math.min(n+1,t.length-1):o=Math.max(n-1,0),I("selectedPlanPath",t[o].path)}async function Se(e){const t=await fetch(e);if(!t.ok)throw new Error(`HTTP ${t.status}`);return t.json()}async function je(){try{const[e,t,n]=await Promise.all([Se("/api/projects"),Se("/api/worktrees"),Se("/api/plans")]);kt(e||[]),Pt(t||[]),Ct(n||[])}catch(e){console.error("Refresh failed:",e)}}let B=null;function nt(e){B&&B.close(),e("connecting"),B=new EventSource("/api/events"),B.addEventListener("connected",()=>{e("ok")}),B.addEventListener("update",t=>{try{const n=JSON.parse(t.data);(n.type==="health"||n.type==="worktrees")&&je()}catch(n){console.error("SSE parse error:",n)}}),B.onerror=()=>{e("error"),setTimeout(()=>nt(e),5e3)}}function qt(){B&&(B.close(),B=null)}const Nt="_tabBar_ll365_1",Rt="_tab_ll365_1",Bt="_active_ll365_26",Lt="_letterKey_ll365_31",ie={tabBar:Nt,tab:Rt,active:Bt,letterKey:Lt};var Ft=A("<nav>"),Mt=A("<button><span>");function Wt(){const e=[{key:"operational",label:"orktrees + PRs",letterKey:"W"},{key:"planning",label:"lans + Issues",letterKey:"P"}];return(()=>{var t=Ft();return b(t,()=>e.map(n=>(()=>{var o=Mt(),r=o.firstChild;return o.$$click=()=>Ae(n.key),b(r,()=>n.letterKey),b(o,()=>n.label,null),m(s=>{var l=`${ie.tab} ${h.currentView===n.key?ie.active:""}`,i=ie.letterKey;return l!==s.e&&d(o,s.e=l),i!==s.t&&d(r,s.t=i),s},{e:void 0,t:void 0}),o})())),m(()=>d(t,ie.tabBar)),t})()}le(["click"]);const Vt="_panel_1ibe1_1",zt="_header_1ibe1_13",Ht="_list_1ibe1_21",Kt="_item_1ibe1_30",Ut="_selected_1ibe1_42",Gt="_count_1ibe1_52",H={panel:Vt,header:zt,list:Ht,item:Kt,selected:Ut,count:Gt};var Jt=A("<div><div>[0] Projects</div><ul tabindex=0 data-panel=project-list>"),Yt=A("<li><span></span><span>");function Xt(){const e=t=>{At(t);const n=h.currentView==="operational"?"worktree-table":"plans-table";V(n),setTimeout(()=>{document.querySelector(`[data-panel="${n}"]`)?.focus()},0)};return(()=>{var t=Jt(),n=t.firstChild,o=n.nextSibling;return b(o,T(xe,{get each(){return h.projects},children:r=>(()=>{var s=Yt(),l=s.firstChild,i=l.nextSibling;return s.$$click=()=>e(r.name),b(l,()=>r.name),b(i,()=>r.count),m(a=>{var u=`${H.item} ${h.selectedProject===r.name?H.selected:""}`,_=H.name,c=H.count;return u!==a.e&&d(s,a.e=u),_!==a.t&&d(l,a.t=_),c!==a.a&&d(i,a.a=c),a},{e:void 0,t:void 0,a:void 0}),s})()})),m(r=>{var s=H.panel,l=H.header,i=H.list;return s!==r.e&&d(t,r.e=s),l!==r.t&&d(n,r.t=l),i!==r.a&&d(o,r.a=i),r},{e:void 0,t:void 0,a:void 0}),t})()}le(["click"]);const Qt="_panel_kz3fu_1",Zt="_header_kz3fu_15",en="_table_kz3fu_23",tn="_tableHeader_kz3fu_31",nn="_sortable_kz3fu_40",sn="_sortAsc_kz3fu_60",rn="_sortDesc_kz3fu_66",on="_tableBody_kz3fu_72",ln="_tableRow_kz3fu_77",an="_selected_kz3fu_95",cn="_colFolder_kz3fu_105",un="_colBranch_kz3fu_106",fn="_colStatus_kz3fu_107",dn="_colPr_kz3fu_108",hn="_statusDirty_kz3fu_110",_n="_statusClean_kz3fu_111",pn="_statusUnpushed_kz3fu_112",vn="_propen_kz3fu_114",bn="_prmerged_kz3fu_115",gn="_prdraft_kz3fu_116",yn="_prclosed_kz3fu_117",$n="_baseIndicator_kz3fu_119",S={panel:Qt,header:Zt,table:en,tableHeader:tn,sortable:nn,sortAsc:sn,sortDesc:rn,tableBody:on,tableRow:ln,selected:an,colFolder:cn,colBranch:un,colStatus:fn,colPr:dn,statusDirty:hn,statusClean:_n,statusUnpushed:pn,propen:vn,prmerged:bn,prdraft:gn,prclosed:yn,baseIndicator:$n};var wn=A("<section><div>[1] Worktrees + PRs</div><div tabindex=0 data-panel=worktree-table><div><span>Folder</span><span>Branch</span><span>Status</span><span>PR</span></div><div>"),Sn=A("<div><span></span><span></span><span></span><span>"),mn=A("<span>BASE"),kn=A("<span>*"),Pn=A("<span>↑"),Cn=A("<span>✓"),An=A("<span>");function jn(e,t,n){const o=[...e],r=n==="asc"?1:-1;if(t==="default"){const s={OPEN:0,DRAFT:1,MERGED:2,CLOSED:3};o.sort((l,i)=>{const a=s[l.prState||""]??4,u=s[i.prState||""]??4;return a!==u?(a-u)*r:l.dirty!==i.dirty?((i.dirty?1:0)-(l.dirty?1:0))*r:l.folder.localeCompare(i.folder)*r})}else if(t==="folder")o.sort((s,l)=>s.folder.localeCompare(l.folder)*r);else if(t==="branch")o.sort((s,l)=>s.branch.localeCompare(l.branch)*r);else if(t==="status")o.sort((s,l)=>{const i=s.dirty?0:s.unpushed>0?1:2,a=l.dirty?0:l.unpushed>0?1:2;return(i-a)*r});else if(t==="pr"){const s={OPEN:0,DRAFT:1,MERGED:2,CLOSED:3};o.sort((l,i)=>{const a=s[l.prState||""]??4,u=s[i.prState||""]??4;return(a-u)*r})}return o}function En(){const e=q(()=>h.worktrees.filter(s=>s.repo===h.selectedProject)),t=q(()=>jn(e(),h.sortColumn,h.sortDirection)),n=s=>{h.sortColumn===s?Re(h.sortDirection==="asc"?"desc":"asc"):(xt(s),Re("asc"))},o=s=>{jt(s),V("worktree-table")},r=s=>h.sortColumn!==s?S.sortable:`${S.sortable} ${h.sortDirection==="asc"?S.sortAsc:S.sortDesc}`;return(()=>{var s=wn(),l=s.firstChild,i=l.nextSibling,a=i.firstChild,u=a.firstChild,_=u.nextSibling,c=_.nextSibling,y=c.nextSibling,g=a.nextSibling;return u.$$click=()=>n("folder"),_.$$click=()=>n("branch"),c.$$click=()=>n("status"),y.$$click=()=>n("pr"),b(g,T(xe,{get each(){return t()},children:f=>(()=>{var $=Sn(),k=$.firstChild,x=k.nextSibling,w=x.nextSibling,P=w.nextSibling;return $.$$click=()=>o(f.folder),b(k,()=>f.folder,null),b(k,(()=>{var v=Y(()=>!!f.base);return()=>v()&&(()=>{var p=mn();return m(()=>d(p,S.baseIndicator)),p})()})(),null),b(x,()=>f.branch),b(w,(()=>{var v=Y(()=>!!f.dirty);return()=>v()&&(()=>{var p=kn();return m(()=>d(p,S.statusDirty)),p})()})(),null),b(w,(()=>{var v=Y(()=>f.unpushed>0);return()=>v()&&(()=>{var p=Pn(),O=p.firstChild;return b(p,()=>f.unpushed,O),m(()=>d(p,S.statusUnpushed)),p})()})(),null),b(w,(()=>{var v=Y(()=>!f.dirty&&f.unpushed===0);return()=>v()&&(()=>{var p=Cn();return m(()=>d(p,S.statusClean)),p})()})(),null),b(P,(()=>{var v=Y(()=>!!f.prState);return()=>v()&&(()=>{var p=An();return b(p,()=>f.prState),m(()=>d(p,S[`pr${f.prState.toLowerCase()}`])),p})()})()),m(v=>{var p=`${S.tableRow} ${h.selectedWorktreeFolder===f.folder?S.selected:""}`,O=S.colFolder,N=S.colBranch,Z=S.colStatus,ee=S.colPr;return p!==v.e&&d($,v.e=p),O!==v.t&&d(k,v.t=O),N!==v.a&&d(x,v.a=N),Z!==v.o&&d(w,v.o=Z),ee!==v.i&&d(P,v.i=ee),v},{e:void 0,t:void 0,a:void 0,o:void 0,i:void 0}),$})()})),m(f=>{var $=S.panel,k=S.header,x=S.table,w=S.tableHeader,P=`${S.colFolder} ${r("folder")}`,v=`${S.colBranch} ${r("branch")}`,p=`${S.colStatus} ${r("status")}`,O=`${S.colPr} ${r("pr")}`,N=S.tableBody;return $!==f.e&&d(s,f.e=$),k!==f.t&&d(l,f.t=k),x!==f.a&&d(i,f.a=x),w!==f.o&&d(a,f.o=w),P!==f.i&&d(u,f.i=P),v!==f.n&&d(_,f.n=v),p!==f.s&&d(c,f.s=p),O!==f.h&&d(y,f.h=O),N!==f.r&&d(g,f.r=N),f},{e:void 0,t:void 0,a:void 0,o:void 0,i:void 0,n:void 0,s:void 0,h:void 0,r:void 0}),s})()}le(["click"]);const xn="_panel_vq0s1_1",Dn="_header_vq0s1_15",On="_table_vq0s1_23",Tn="_tableHeader_vq0s1_31",In="_sortable_vq0s1_40",qn="_sortAsc_vq0s1_60",Nn="_sortDesc_vq0s1_66",Rn="_tableBody_vq0s1_72",Bn="_tableRow_vq0s1_77",Ln="_selected_vq0s1_95",Fn="_colTitle_vq0s1_105",Mn="_colProject_vq0s1_106",Wn="_colStatus_vq0s1_107",Vn="_colIssue_vq0s1_108",zn="_statusdraft_vq0s1_110",Hn="_statusactive_vq0s1_111",Kn="_statusdone_vq0s1_112",Un="_statusarchived_vq0s1_113",Gn="_statuscompleted_vq0s1_114",Jn="_statusplanned_vq0s1_115",j={panel:xn,header:Dn,table:On,tableHeader:Tn,sortable:In,sortAsc:qn,sortDesc:Nn,tableBody:Rn,tableRow:Bn,selected:Ln,colTitle:Fn,colProject:Mn,colStatus:Wn,colIssue:Vn,statusdraft:zn,statusactive:Hn,statusdone:Kn,statusarchived:Un,statuscompleted:Gn,statusplanned:Jn};var Yn=A("<section><div>[1] Plans + Issues</div><div tabindex=0 data-panel=plans-table><div><span>Title</span><span>Project</span><span>Status</span><span>Issue</span></div><div>"),Xn=A("<div><span></span><span></span><span><span></span></span><span>");function Qn(e,t,n){const o=[...e],r=n==="asc"?1:-1;if(t==="default"){const s={draft:0,active:0,done:1,archived:2};o.sort((l,i)=>{const a=s[l.status]??3,u=s[i.status]??3;return a!==u?(a-u)*r:l.title.localeCompare(i.title)*r})}else t==="title"?o.sort((s,l)=>s.title.localeCompare(l.title)*r):t==="project"?o.sort((s,l)=>s.project.localeCompare(l.project)*r):t==="status"?o.sort((s,l)=>(s.status||"").localeCompare(l.status||"")*r):t==="issue"&&o.sort((s,l)=>((s.issue||0)-(l.issue||0))*r);return o}function Zn(){const e=q(()=>h.plans.filter(s=>s.project===h.selectedProject)),t=q(()=>Qn(e(),h.planSortColumn,h.planSortDirection)),n=s=>{h.planSortColumn===s?Be(h.planSortDirection==="asc"?"desc":"asc"):(Dt(s),Be("asc"))},o=s=>{Et(s),V("plans-table")},r=s=>h.planSortColumn!==s?j.sortable:`${j.sortable} ${h.planSortDirection==="asc"?j.sortAsc:j.sortDesc}`;return(()=>{var s=Yn(),l=s.firstChild,i=l.nextSibling,a=i.firstChild,u=a.firstChild,_=u.nextSibling,c=_.nextSibling,y=c.nextSibling,g=a.nextSibling;return u.$$click=()=>n("title"),_.$$click=()=>n("project"),c.$$click=()=>n("status"),y.$$click=()=>n("issue"),b(g,T(xe,{get each(){return t()},children:f=>(()=>{var $=Xn(),k=$.firstChild,x=k.nextSibling,w=x.nextSibling,P=w.firstChild,v=w.nextSibling;return $.$$click=()=>o(f.path),b(k,()=>f.title),b(x,()=>f.project),b(P,()=>f.status||"draft"),b(v,(()=>{var p=Y(()=>!!f.issue);return()=>p()&&`#${f.issue}`})()),m(p=>{var O=`${j.tableRow} ${h.selectedPlanPath===f.path?j.selected:""}`,N=j.colTitle,Z=j.colProject,ee=j.colStatus,De=j[`status${f.status}`],Oe=j.colIssue;return O!==p.e&&d($,p.e=O),N!==p.t&&d(k,p.t=N),Z!==p.a&&d(x,p.a=Z),ee!==p.o&&d(w,p.o=ee),De!==p.i&&d(P,p.i=De),Oe!==p.n&&d(v,p.n=Oe),p},{e:void 0,t:void 0,a:void 0,o:void 0,i:void 0,n:void 0}),$})()})),m(f=>{var $=j.panel,k=j.header,x=j.table,w=j.tableHeader,P=`${j.colTitle} ${r("title")}`,v=`${j.colProject} ${r("project")}`,p=`${j.colStatus} ${r("status")}`,O=`${j.colIssue} ${r("issue")}`,N=j.tableBody;return $!==f.e&&d(s,f.e=$),k!==f.t&&d(l,f.t=k),x!==f.a&&d(i,f.a=x),w!==f.o&&d(a,f.o=w),P!==f.i&&d(u,f.i=P),v!==f.n&&d(_,f.n=v),p!==f.s&&d(c,f.s=p),O!==f.h&&d(y,f.h=O),N!==f.r&&d(g,f.r=N),f},{e:void 0,t:void 0,a:void 0,o:void 0,i:void 0,n:void 0,s:void 0,h:void 0,r:void 0}),s})()}le(["click"]);const es="_section_1rmen_1",ts="_header_1rmen_5",ns="_panel_1rmen_15",ss="_empty_1rmen_29",rs="_content_1rmen_33",os="_row_1rmen_39",ls="_label_1rmen_43",is="_value_1rmen_49",M={section:es,header:ts,panel:ns,empty:ss,content:rs,row:os,label:ls,value:is};var as=A("<div>"),cs=A("<section><div>[2] Details</div><div tabindex=0 data-panel=details>"),us=A("<span>Select an item to view details"),fs=A("<div><span></span><span>");function ds(){const e=q(()=>Ot()),t=q(()=>Tt()),n=q(()=>{const s=e();if(!s)return null;const l=[{label:"Folder:",value:s.folder},{label:"Repo:",value:s.repo},{label:"Branch:",value:s.branch},{label:"Base:",value:s.base?"Yes":"No"}];s.purpose&&l.push({label:"Purpose:",value:s.purpose}),s.status&&l.push({label:"Status:",value:s.status});const i=[];return s.dirty&&i.push("Uncommitted changes"),s.unpushed>0&&i.push(`${s.unpushed} unpushed`),s.prState&&i.push(`PR: ${s.prState}`),i.length>0&&l.push({label:"Health:",value:i.join(", ")}),l}),o=q(()=>{const s=t();if(!s)return null;const l=[{label:"Title:",value:s.title},{label:"Project:",value:s.project},{label:"Status:",value:s.status||"draft"},{label:"Path:",value:s.path}];return s.issue&&l.push({label:"Issue:",value:`#${s.issue}`}),s.priority&&l.push({label:"Priority:",value:String(s.priority)}),l}),r=q(()=>h.currentView==="operational"?n():o());return(()=>{var s=cs(),l=s.firstChild,i=l.nextSibling;return b(i,T(pe,{get when(){return r()},get fallback(){return(()=>{var a=us();return m(()=>d(a,M.empty)),a})()},get children(){var a=as();return b(a,()=>r().map(u=>(()=>{var _=fs(),c=_.firstChild,y=c.nextSibling;return b(c,()=>u.label),b(y,()=>u.value),m(g=>{var f=M.row,$=M.label,k=M.value;return f!==g.e&&d(_,g.e=f),$!==g.t&&d(c,g.t=$),k!==g.a&&d(y,g.a=k),g},{e:void 0,t:void 0,a:void 0}),_})())),m(()=>d(a,M.content)),a}})),m(a=>{var u=M.section,_=M.header,c=M.panel;return u!==a.e&&d(s,a.e=u),_!==a.t&&d(l,a.t=_),c!==a.a&&d(i,a.a=c),a},{e:void 0,t:void 0,a:void 0}),s})()}const hs="_footer_u37gh_1",_s="_status_u37gh_10",ps="_statusOk_u37gh_15",vs="_statusError_u37gh_19",bs="_statusConnecting_u37gh_23",gs="_help_u37gh_27",J={footer:hs,status:_s,statusOk:ps,statusError:vs,statusConnecting:bs,help:gs};var ys=A("<footer><span></span><span>Press <kbd>?</kbd> for help");function $s(e){const t=()=>{switch(e.connectionStatus){case"ok":return"Connected";case"error":return"Disconnected";case"connecting":return"Connecting..."}},n=()=>{switch(e.connectionStatus){case"ok":return J.statusOk;case"error":return J.statusError;case"connecting":return J.statusConnecting}};return(()=>{var o=ys(),r=o.firstChild,s=r.nextSibling;return b(r,t),m(l=>{var i=J.footer,a=`${J.status} ${n()}`,u=J.help;return i!==l.e&&d(o,l.e=i),a!==l.t&&d(r,l.t=a),u!==l.a&&d(s,l.a=u),l},{e:void 0,t:void 0,a:void 0}),o})()}const ws="_overlay_wdxgf_1",Ss="_modal_wdxgf_14",ms="_header_wdxgf_22",ks="_close_wdxgf_34",Ps="_content_wdxgf_48",Cs="_row_wdxgf_55",As="_key_wdxgf_61",js="_desc_wdxgf_72",W={overlay:ws,modal:Ss,header:ms,close:ks,content:Ps,row:Cs,key:As,desc:js};var Es=A("<div><div><div><span>Keyboard Shortcuts</span><button>×</button></div><div>"),xs=A("<div><kbd></kbd><span>");function Ds(e){const t=[{key:"0/1/2",desc:"Focus panel (Projects/Main/Details)"},{key:"w",desc:"Worktrees + PRs view"},{key:"p",desc:"Plans + Issues view"},{key:"j/k ↑/↓",desc:"Navigate up/down"},{key:"h/l ←/→",desc:"Navigate left/right"},{key:"r",desc:"Refresh data"},{key:"?",desc:"Toggle help"},{key:"Esc",desc:"Close modal"}];return T(pe,{get when(){return e.open},get children(){var n=Es(),o=n.firstChild,r=o.firstChild,s=r.firstChild,l=s.nextSibling,i=r.nextSibling;return qe(n,"click",e.onClose),o.$$click=a=>a.stopPropagation(),qe(l,"click",e.onClose),b(i,()=>t.map(a=>(()=>{var u=xs(),_=u.firstChild,c=_.nextSibling;return b(_,()=>a.key),b(c,()=>a.desc),m(y=>{var g=W.row,f=W.key,$=W.desc;return g!==y.e&&d(u,y.e=g),f!==y.t&&d(_,y.t=f),$!==y.a&&d(c,y.a=$),y},{e:void 0,t:void 0,a:void 0}),u})())),m(a=>{var u=W.overlay,_=W.modal,c=W.header,y=W.close,g=W.content;return u!==a.e&&d(n,a.e=u),_!==a.t&&d(o,a.t=_),c!==a.a&&d(r,a.a=c),y!==a.o&&d(l,a.o=y),g!==a.i&&d(i,a.i=g),a},{e:void 0,t:void 0,a:void 0,o:void 0,i:void 0}),n}})}le(["click"]);const Os="_app_1im25_1",Ts="_main_1im25_8",Is="_sidebar_1im25_14",qs="_content_1im25_22",ae={app:Os,main:Ts,sidebar:Is,content:qs};var Ns=A("<div><div><aside></aside><div>");function Rs(){const[e,t]=de("connecting"),[n,o]=de(!1);return it(()=>{je(),nt(t);const r=s=>{if(!(s.target instanceof HTMLInputElement||s.target instanceof HTMLTextAreaElement))switch(s.key){case"w":Ae("operational");break;case"p":Ae("planning");break;case"0":V("project-list"),document.querySelector('[data-panel="project-list"]')?.focus();break;case"1":const l=h.currentView==="operational"?"worktree-table":"plans-table";V(l),document.querySelector(`[data-panel="${l}"]`)?.focus();break;case"2":V("details"),document.querySelector('[data-panel="details"]')?.focus();break;case"j":case"ArrowDown":s.preventDefault(),h.focusedPanel==="project-list"?(Le("down"),document.querySelector('[data-panel="project-list"]')?.focus()):h.focusedPanel==="worktree-table"?(Fe("down"),document.querySelector('[data-panel="worktree-table"]')?.focus()):h.focusedPanel==="plans-table"&&(Me("down"),document.querySelector('[data-panel="plans-table"]')?.focus());break;case"k":case"ArrowUp":s.preventDefault(),h.focusedPanel==="project-list"?(Le("up"),document.querySelector('[data-panel="project-list"]')?.focus()):h.focusedPanel==="worktree-table"?(Fe("up"),document.querySelector('[data-panel="worktree-table"]')?.focus()):h.focusedPanel==="plans-table"&&(Me("up"),document.querySelector('[data-panel="plans-table"]')?.focus());break;case"l":case"ArrowRight":if(s.preventDefault(),h.focusedPanel==="project-list"){const i=h.currentView==="operational"?"worktree-table":"plans-table";V(i)}break;case"h":case"ArrowLeft":s.preventDefault(),V("project-list");break;case"r":je();break;case"?":o(i=>!i);break;case"Escape":o(!1);break}};document.addEventListener("keydown",r),He(()=>{document.removeEventListener("keydown",r),qt()})}),(()=>{var r=Ns(),s=r.firstChild,l=s.firstChild,i=l.nextSibling;return b(r,T(Wt,{}),s),b(l,T(Xt,{})),b(i,T(pe,{get when(){return h.currentView==="operational"},get children(){return T(En,{})}}),null),b(i,T(pe,{get when(){return h.currentView==="planning"},get children(){return T(Zn,{})}}),null),b(i,T(ds,{}),null),b(r,T($s,{get connectionStatus(){return e()},onShowHelp:()=>o(!0)}),null),b(r,T(Ds,{get open(){return n()},onClose:()=>o(!1)}),null),m(a=>{var u=ae.app,_=ae.main,c=ae.sidebar,y=ae.content;return u!==a.e&&d(r,a.e=u),_!==a.t&&d(s,a.t=_),c!==a.a&&d(l,a.a=c),y!==a.o&&d(i,a.o=y),a},{e:void 0,t:void 0,a:void 0,o:void 0}),r})()}const Bs=document.getElementById("root");vt(()=>T(Rs,{}),Bs);
