--- web/playwright.config.js
@@ -1,28 +1,0 @@
-// @ts-check
-const { defineConfig, devices } = require('@playwright/test');
-
-module.exports = defineConfig({
-  testDir: './tests',
-  fullyParallel: true,
-  forbidOnly: !!process.env.CI,
-  retries: process.env.CI ? 2 : 0,
-  workers: process.env.CI ? 1 : undefined,
-  reporter: 'html',
-  use: {
-    baseURL: 'http://localhost:8374',
-    trace: 'on-first-retry',
-    screenshot: 'only-on-failure',
-  },
-  projects: [
-    {
-      name: 'chromium',
-      use: { ...devices['Desktop Chrome'] },
-    },
-  ],
-  webServer: {
-    command: 'cd .. && go run ./cmd/bearing daemon start --foreground',
-    url: 'http://localhost:8374',
-    reuseExistingServer: true,
-    timeout: 30000,
-  },
-});
